vrouter: Changes required for vrouter to build after update from master.
Some function definitions have changed. Additionaly the missing prototypes
warning warning have to be disabled, because the sandesh generated code is
not buildable with this flag enabled as it gives no prototypes in code.

Bug has been reported for missing vhost_remove_xconnect function and
for providing workaround and enabling back the missing prototypes
warning.